requirements.txt
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
google-genai>=0.2.0
pillow>=10.0.0
